Commit - Determination of Null Vs Non Null
Author: Steven Greulich
Purpose of Commit: The purpose of this commit was to determine  if the AssessJob columns that have NaN should be removed or not from the dataset.  After consideration, they will not be filled in with values, but would just be removed from the dataset.  However, they will be mentioned in the report to ensure a full picture of the data counts