Commit: Update of Clean_and_plot method
Author: Steven Greulich
Purpose of Commit: The purpose of this commit was to update the Clean_and_plot method.  The original issue was that the bar chart was only printing out the first letter of each array for the x axis.  The method has been changed so that it now correctly prints out the full technology name as expected